only new core

+++ cpp/splinepy/splines/create_bspline.cpp
@@ -8,7 +8,7 @@ std::shared_ptr<SplinepyBase>
 splinepy::splines::SplinepyBase::SplinepyCreateBSpline(
     const int para_dim,
     const int dim,
-    const double* degrees,
+    const int* degrees,
     const std::vector<std::vector<double>>* knot_vectors,
     const double* control_points) {
   switch (para_dim) {
